fixed some issues with linker
