--- AnalisysAlgorithms_Lab1_Sort/AnalisysAlgorithms_Lab1_Sort.sln
@@ -1,38 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "SortUnitTest", "SortUnitTest\SortUnitTest.vcxproj", "{7B09ADE6-FD1A-4D58-97F5-FA2806EBF596}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Sort", "Sort\Sort.vcxproj", "{9DF40486-311C-49DA-915E-4B80CCAE3D17}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7B09ADE6-FD1A-4D58-97F5-FA2806EBF596}.Debug|x64.ActiveCfg = Debug|x64
-		{7B09ADE6-FD1A-4D58-97F5-FA2806EBF596}.Debug|x64.Build.0 = Debug|x64
-		{7B09ADE6-FD1A-4D58-97F5-FA2806EBF596}.Debug|x86.ActiveCfg = Debug|Win32
-		{7B09ADE6-FD1A-4D58-97F5-FA2806EBF596}.Debug|x86.Build.0 = Debug|Win32
-		{7B09ADE6-FD1A-4D58-97F5-FA2806EBF596}.Release|x64.ActiveCfg = Release|x64
-		{7B09ADE6-FD1A-4D58-97F5-FA2806EBF596}.Release|x64.Build.0 = Release|x64
-		{7B09ADE6-FD1A-4D58-97F5-FA2806EBF596}.Release|x86.ActiveCfg = Release|Win32
-		{7B09ADE6-FD1A-4D58-97F5-FA2806EBF596}.Release|x86.Build.0 = Release|Win32
-		{9DF40486-311C-49DA-915E-4B80CCAE3D17}.Debug|x64.ActiveCfg = Debug|x64
-		{9DF40486-311C-49DA-915E-4B80CCAE3D17}.Debug|x64.Build.0 = Debug|x64
-		{9DF40486-311C-49DA-915E-4B80CCAE3D17}.Debug|x86.ActiveCfg = Debug|Win32
-		{9DF40486-311C-49DA-915E-4B80CCAE3D17}.Debug|x86.Build.0 = Debug|Win32
-		{9DF40486-311C-49DA-915E-4B80CCAE3D17}.Release|x64.ActiveCfg = Release|x64
-		{9DF40486-311C-49DA-915E-4B80CCAE3D17}.Release|x64.Build.0 = Release|x64
-		{9DF40486-311C-49DA-915E-4B80CCAE3D17}.Release|x86.ActiveCfg = Release|Win32
-		{9DF40486-311C-49DA-915E-4B80CCAE3D17}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
