--- AnalisysAlgorithms_Lab1_Sort/Sort/Sort.h
@@ -1,16 +1,0 @@
-//#pragma once
-//
-////template<typename T>
-////bool comp(const T& a, const T& b);
-//
-//template<typename T>
-//void swap(T& a, T& b);
-//
-//template <typename T, typename Compare>
-//void sort(T *first, T *last, Compare comp);
-//
-//template <typename T, typename Compare>
-//void insertion_sort(T *vec, T *last, Compare comp);
-//
-//template <typename T, typename Compare>
-//void partition(T *first, T *last, Compare comp);
