--- AnalisysAlgorithms_Lab1_Sort/SortUnitTest/stdafx.cpp
@@ -1,8 +1,0 @@
-// stdafx.cpp: файл исходного кода, содержащий стандартные включаемые элементы
-// SortUnitTest.pch будет предкомпилированным заголовком
-// stdafx.obj будет содержать предварительно откомпилированные сведения о типе
-
-#include "stdafx.h"
-
-// TODO: Установите ссылки на любые требующиеся дополнительные заголовки в файле STDAFX.H
-// , а не в данном файле
