--- AnalisysAlgorithms_Lab1_Sort/SortUnitTest/targetver.h
@@ -1,8 +1,0 @@
-#pragma once
-
-// Включение SDKDDKVer.h обеспечивает определение самой последней доступной платформы Windows.
-
-// Если требуется выполнить сборку приложения для предыдущей версии Windows, включите WinSDKVer.h и
-// задайте для макроса _WIN32_WINNT платформу, которая должна поддерживаться, прежде чем включать SDKDDKVer.h.
-
-#include <SDKDDKVer.h>
