--- AnalisysAlgorithms_Lab1_Sort/SortUnitTest/unittest1.cpp
@@ -1,19 +1,0 @@
-#include "stdafx.h"
-#include "CppUnitTest.h"
-#include "..\Sort\Sort.h"
-
-using namespace Microsoft::VisualStudio::CppUnitTestFramework;
-
-namespace SortUnitTest
-{		
-	TEST_CLASS(UnitTest1)
-	{
-	public:
-		
-		TEST_METHOD(TestMethod1)
-		{
-			
-		}
-
-	};
-}
