+++ Sort/Sort.sln
@@ -1,0 +1,38 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.25420.1
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Sort", "Sort\Sort.vcxproj", "{B88B1EC6-9EB3-4118-B43F-795EC5B6F2A1}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "UnitTest1", "UnitTest1\UnitTest1.vcxproj", "{0538F45F-3F16-4436-B089-99537B04E8FC}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B88B1EC6-9EB3-4118-B43F-795EC5B6F2A1}.Debug|x64.ActiveCfg = Debug|x64
+		{B88B1EC6-9EB3-4118-B43F-795EC5B6F2A1}.Debug|x64.Build.0 = Debug|x64
+		{B88B1EC6-9EB3-4118-B43F-795EC5B6F2A1}.Debug|x86.ActiveCfg = Debug|Win32
+		{B88B1EC6-9EB3-4118-B43F-795EC5B6F2A1}.Debug|x86.Build.0 = Debug|Win32
+		{B88B1EC6-9EB3-4118-B43F-795EC5B6F2A1}.Release|x64.ActiveCfg = Release|x64
+		{B88B1EC6-9EB3-4118-B43F-795EC5B6F2A1}.Release|x64.Build.0 = Release|x64
+		{B88B1EC6-9EB3-4118-B43F-795EC5B6F2A1}.Release|x86.ActiveCfg = Release|Win32
+		{B88B1EC6-9EB3-4118-B43F-795EC5B6F2A1}.Release|x86.Build.0 = Release|Win32
+		{0538F45F-3F16-4436-B089-99537B04E8FC}.Debug|x64.ActiveCfg = Debug|x64
+		{0538F45F-3F16-4436-B089-99537B04E8FC}.Debug|x64.Build.0 = Debug|x64
+		{0538F45F-3F16-4436-B089-99537B04E8FC}.Debug|x86.ActiveCfg = Debug|Win32
+		{0538F45F-3F16-4436-B089-99537B04E8FC}.Debug|x86.Build.0 = Debug|Win32
+		{0538F45F-3F16-4436-B089-99537B04E8FC}.Release|x64.ActiveCfg = Release|x64
+		{0538F45F-3F16-4436-B089-99537B04E8FC}.Release|x64.Build.0 = Release|x64
+		{0538F45F-3F16-4436-B089-99537B04E8FC}.Release|x86.ActiveCfg = Release|Win32
+		{0538F45F-3F16-4436-B089-99537B04E8FC}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
