+++ Sort/Sort/ReadMe.txt
@@ -1,0 +1,30 @@
+﻿========================================================================
+    КОНСОЛЬНОЕ ПРИЛОЖЕНИЕ. Обзор проекта Sort
+========================================================================
+
+Это приложение Sort создано автоматически с помощью мастера приложений.
+
+В этом файле представлена сводка содержимого всех файлов, входящих в состав приложения Sort.
+
+
+Sort.vcxproj
+    Это основной файл проекта VC++, создаваемый с помощью мастера приложений. Он содержит данные о версии языка Visual C++, использованной для создания файла, а также сведения о платформах, конфигурациях и функциях проекта, выбранных с помощью мастера приложений.
+
+Sort.vcxproj.filters
+    Это файл фильтров для проектов VC++, созданный с помощью мастера приложений. Он содержит сведения о сопоставлениях между файлами в вашем проекте и фильтрами. Эти сопоставления используются в среде IDE для группировки файлов с одинаковыми расширениями в одном узле (например CPP-файлы сопоставляются с фильтром "Исходные файлы").
+
+Sort.cpp
+    Это основной исходный файл приложения.
+
+/////////////////////////////////////////////////////////////////////////////
+Другие стандартные файлы:
+
+StdAfx.h, StdAfx.cpp
+    Эти файлы используются для построения файла предкомпилированного заголовка (PCH) с именем Sort.pch и файла предкомпилированных типов с именем StdAfx.obj.
+
+/////////////////////////////////////////////////////////////////////////////
+Прочие примечания.
+
+С помощью комментариев «TODO:» в мастере приложений обозначаются фрагменты исходного кода, которые необходимо дополнить или изменить.
+
+/////////////////////////////////////////////////////////////////////////////
