+++ Sort/Sort/Sort.h
@@ -1,0 +1,20 @@
+#pragma once
+#include "Sort.cpp"
+
+
+////namespace Sorting {
+//	//template<typename T>
+//	//bool comp(const T& a, const T& b);
+//
+//	template<typename T>
+//	void swap(T a, T b);
+//
+//	template <typename T, typename Compare>
+//	void sort(T *first, T *last, Compare comp);
+//
+//	template <typename T, typename Compare>
+//	void insertion_sort(T *vec, T *last, Compare comp);
+//
+//	template <typename T, typename Compare>
+//	void partition(T *first, T *last, Compare comp);
+////}
