+++ Sort/Sort/stdafx.cpp
@@ -1,0 +1,8 @@
+// stdafx.cpp: исходный файл, содержащий только стандартные включаемые модули
+// Sort.pch будет предкомпилированным заголовком
+// stdafx.obj будет содержать предварительно откомпилированные сведения о типе
+
+#include "stdafx.h"
+
+// TODO: Установите ссылки на любые требующиеся дополнительные заголовки в файле STDAFX.H
+// , а не в данном файле
