+++ Sort/Sort/stdafx.h
@@ -1,0 +1,15 @@
+// stdafx.h: включаемый файл дл€ стандартных системных включаемых файлов
+// или включаемых файлов дл€ конкретного проекта, которые часто используютс€, но
+// не часто измен€ютс€
+//
+
+#pragma once
+
+#include "targetver.h"
+
+#include <stdio.h>
+#include <tchar.h>
+
+
+
+// TODO: ”становите здесь ссылки на дополнительные заголовки, требующиес€ дл€ программы
