+++ Sort/UnitTest1/stdafx.cpp
@@ -1,0 +1,8 @@
+// stdafx.cpp: файл исходного кода, содержащий стандартные включаемые элементы
+// UnitTest1.pch будет предкомпилированным заголовком
+// stdafx.obj будет содержать предварительно откомпилированные сведения о типе
+
+#include "stdafx.h"
+
+// TODO: Установите ссылки на любые требующиеся дополнительные заголовки в файле STDAFX.H
+// , а не в данном файле
