+++ Sort/UnitTest1/stdafx.h
@@ -1,0 +1,13 @@
+// stdafx.h: включаемый файл для стандартных системных включаемых файлов,
+// или включаемых файлов для конкретного проекта, которые часто используются, но
+// не часто изменяются
+//
+
+#pragma once
+
+#include "targetver.h"
+
+// Заголовки CppUnitTest
+#include "CppUnitTest.h"
+
+// TODO: Установите здесь ссылки на дополнительные заголовки, требующиеся для программы
